re-organized into projects and cases, allow for organized running of many different run cases (aka scenarios/alternatives)
