+++ .env.example
@@ -1,34 +1,38 @@
-# Copy to .env and adjust for your machine.  Everything machine-specific lives
-# here, so config.yaml stays the model recipe and nothing else.
-#
-# The project config reads these as {env:NAME}.  An unset variable is an error
-# naming it, not a silent empty string.
+# Copy to .env and adjust for this machine.  Machine-specific things only, so
+# config.yaml stays the model recipe.  The config reads these as {env:NAME}; an
+# unset one is an error naming it, never a silent empty string.
 
-# --- required: where a run reads from, and where it happens ------------------
+# --- required ----------------------------------------------------------------
 
-# The MTC model data share, and nothing more of the path than that: which network,
-# which land use, which previous run are written out in config.yaml where they can be
-# read.  Use the drive letter if it is mapped on this machine, the UNC path if not --
-# that difference is exactly what belongs in .env.
-#   M: = \models.ad.mtc.ca.gov\data\models
+# The model data share, and no more of the path than that -- which network, which
+# land use, which previous run are written out in config.yaml where they can be
+# reviewed.  Drive letter where it is mapped, UNC where it is not; that difference
+# is the whole reason this is here.
 TM1_M_DRIVE=//models.ad.mtc.ca.gov/data/models
 
-# Where the run happens.  Needs room for the full model output (~100 GB) and
-# should be a local disk -- Cube's cluster is chatty and a network path is slow.
-TM1_PROJ_DIR=E:/Tests/base_2023_ctramp
+# Where runs go, as {project}/{case}-{NNN} beneath this.  Local disk (~100 GB per
+# run, and Cube's cluster is too chatty for a network path), and keep it SHORT --
+# Cube and the Java stack are not long-path aware and a run nests ~160 characters.
+TM1_RUNS_ROOT=E:/runs
 
-# --- required: what the legacy model code needs to find on this machine ------
-
-# BuildTransitNetworks.job calls gawk itself -- to derive transit egress links and to
-# pick park-and-ride lots -- so gawk has to be on PATH when Cube runs it, as it was in
-# RunModel.bat's shell.  Windows has no awk; Git for Windows bundles one at usr/bin.
-# A wrong path does not make the job exit non-zero -- the shell redirect creates its
-# output files, empty -- so build_transit_networks declares them under `verify:` and
-# fails there rather than three steps later.
+# gawk, which BuildTransitNetworks.job shells out to.  Windows has none; Git for
+# Windows bundles one.  A wrong path here does not fail loudly -- the shell redirect
+# creates the output files, empty -- so the step verifies them and fails there.
 TM1_GAWK_DIR=C:/Program Files/Git/usr/bin
 
-# --- optional: only if these are in non-standard locations -------------------
-# VoyagerFileAccess DLLs are bundled -- no config needed for TPP->OMX.
+# --- optional ----------------------------------------------------------------
 
-# RUNTPP_PATH=C:\Program Files\Citilabs\CubeVoyager
-# R_HOME=C:\Program Files\R\R-4.2.1
+# The three Cube files this needs.  Shown here are the installer defaults, already
+# set in code -- name one only when this machine differs.  Their directories go on
+# PATH and Java's library path automatically, so a Bentley OpenPaths install is just
+# TM1_RUNTPP pointing there, not a separate setting.
+
+# TM1_RUNTPP=C:\Program Files\Citilabs\CubeVoyager\runtpp.exe
+# TM1_CLUSTER=C:\Program Files\Citilabs\CubeVoyager\Cluster.exe
+# TM1_VOYAGER_DLL=C:\Program Files\Citilabs\VoyagerFileAPI\VoyagerFileAccess.dll
+
+# Slack.  Either the URL, or a file holding it so a shared webhook stays off every
+# machine.  With neither, runs log their milestones and send nothing.  How much is
+# sent is `slack:` in the project config, not here.
+TM1_SLACK_WEBHOOK_FILE=//models.ad.mtc.ca.gov/data/models/Software/Slack/TravelModel_SlackWebhook.txt
+# SLACK_WEBHOOK_URL=https://hooks.slack.com/services/...
