+++ projects/ctramp_2023/cases.yaml
@@ -1,0 +1,92 @@
+# Cases for ctramp_2023.
+#
+# A case is a named set of overrides into config.yaml, addressed by where the
+# value lives:
+#
+#   m_drive: "X:/models"                              a top-level key
+#   env.MODEL_YEAR: 2035                              inside a top-level mapping
+#   env.EN7: ENABLED                                  ... and its siblings survive
+#   copy_inputs.input_landuse.from: "M:/.../landuse"  inside a step
+#   iterate.count: 1                                  the loop's own key
+#   warmstart.hwy_assign.cluster_nodes: 24            a step, in one block
+#
+# The address must already exist, values replace rather than merge, a step name
+# appearing in both `warmstart:` and `iterate:` must say which, and `steps` itself
+# is not addressable -- a case varies values, it never changes the pipeline.
+#
+# Three ways to declare them, all expanding to the same thing:
+#
+#   cases:    written out, one entry per run
+#   ladder:   cumulative -- rung k applies rungs 1..k, so the diff between
+#             adjacent rungs isolates the one intervention that was added
+#   matrix:   the cross product of named axes, minus `exclude:` combinations
+#
+# `tm1 cases ctramp_2023` prints what this file expands to and checks every
+# address against config.yaml.
+#
+# IDs are SERIES-TOKENS-YEAR, uppercase, and stable forever: the ID names the run
+# directory, so renaming a case makes it an unrun one.
+
+cases:
+
+  # No overrides: config.yaml exactly as written.
+  PARITY-2023:
+    description: >
+      RunModel.bat parity: adopted 2023 networks and land use, no strategies
+      applied.  The run the harness is measured against.
+
+# ── The other two pathways, for when this project grows past one case ──────────────────
+# Both expand to the same thing `cases:` does, so the rules above apply unchanged.
+# Uncomment, adjust, and check with `tm1 cases ctramp_2023`.
+
+# ladder: cumulative un-ablation.  Start from the year's land use and networks with no
+# strategies, then switch them on one at a time until the last rung *is* the full
+# Blueprint.  The difference between adjacent rungs is that one strategy's contribution,
+# which is the question these runs exist to answer.
+#
+# ladder:
+#   - id: "BP-{n}-{rung}-2035"
+#     description: 2035 Blueprint, un-ablated one strategy at a time.
+#     env.MODEL_YEAR: 2035
+#     rungs:
+#       - NONE:
+#           description: 2035 land use and adopted networks, no strategies.
+#       - TELE:
+#           description: + telecommute.
+#           env.EN7: ENABLED
+#       - TRNF:
+#           description: + transit frequency.
+#           copy_inputs.input_trn.from: "{m_drive}/.../net_2035_Blueprint/trn"
+#       - CORD:
+#           description: + cordon pricing.
+#           set_tolls.job: "variants/SetTolls_cordon.job"
+#
+#   gives  BP-01-NONE-2035  BP-02-TELE-2035  BP-03-TRNF-2035  BP-04-CORD-2035
+#   BP-04 carries all four, so it is the full Blueprint; BP-01 is the do-nothing floor.
+#   Inserting a rung renumbers everything after it -- correctly, since it changes what
+#   every later rung means.
+
+# matrix: the cross product of named axes, for a sweep.  Give every axis point a
+# description -- a case's prose is assembled from them, and in a sweep of thirty the
+# ID is otherwise all you have to go on.
+#
+# matrix:
+#   - id: "A1-{tolls}-{landuse}"
+#     axes:
+#       tolls:
+#         NOTL:
+#           description: No new tolls.
+#         CORD:
+#           description: Cordon pricing.
+#           set_tolls.job: "variants/SetTolls_cordon.job"
+#       landuse:
+#         ADPT:
+#           description: Adopted land use.
+#         JHBL:
+#           description: Jobs-housing balance.
+#           env.MODEL_YEAR: 2035
+#     exclude:
+#       - {tolls: CORD, landuse: JHBL}     # combinations that make no sense
+#
+#   gives  A1-NOTL-ADPT, A1-NOTL-JHBL, A1-CORD-ADPT
+#   (an exclude rule may name a subset of the axes, so it survives a new axis)
