+++ src/tm1/assignment/cube/__init__.py
@@ -1,6 +1,6 @@
 """Faithful Cube Voyager assignment backend (run as-is via ``runtpp``).
 
-- :mod:`~tm1.assignment.cube.runner` -- run Cube ``.job`` scripts over
+- :mod:`cube.job` -- run Cube ``.job`` scripts over
   SSH/schtasks, including the local Cube Cluster.
 - :mod:`~tm1.assignment.cube.highway` -- HwyAssign, the feedback block, and
   highway skims.
@@ -10,6 +10,6 @@
   global iteration.
 """
 
-from tm1.assignment.cube.runner import CubeJobError, run_cube_job
+from cube.job import CubeJobError, run_cube_job
 
 __all__ = ["CubeJobError", "run_cube_job"]
